Moved buddies loading logic to a sync helper
